set up for wallet-initiated testing
