+++ vercre-issuer/tests/wallet_initiated.rs
@@ -1,19 +1,22 @@
-//! Authorization Code Flow Tests
+//! Wallet-initiated Tests
 
 mod utils;
 mod wallet;
 
 use rstest::rstest;
+use utils::{provider, Issuance};
 use vercre_openid::CredentialFormat;
-use vercre_test_utils::issuer;
+use vercre_test_utils::issuer::Provider;
 
 #[rstest]
-async fn authorization() {
+#[case(Issuance::Immediate)]
+#[case(Issuance::Deferred)]
+async fn issuance(provider: Provider, #[case] issue: Issuance) {
     vercre_test_utils::init_tracer();
-    snapshot!("authorization");
+    snapshot!("wallet:{issue}");
 
     let wallet = wallet::Wallet {
-        provider: issuer::Provider::new(),
+        provider,
         format: CredentialFormat::JwtVcJson,
         ..Default::default()
     };
